Separate return value from error code in gen_irq_code*().

+++ uspace/lib/c/generic/event.c
@@ -44,12 +44,12 @@
  * @param evno    Event type to subscribe.
  * @param imethod Use this interface and method for notifying me.
  *
- * @return Value returned by the kernel.
+ * @return Error code returned by the kernel.
  *
  */
 int ipc_event_subscribe(event_type_t evno, sysarg_t imethod)
 {
-	return __SYSCALL2(SYS_IPC_EVENT_SUBSCRIBE, (sysarg_t) evno,
+	return (int) __SYSCALL2(SYS_IPC_EVENT_SUBSCRIBE, (sysarg_t) evno,
 	    (sysarg_t) imethod);
 }
 
@@ -58,12 +58,12 @@ int ipc_event_subscribe(event_type_t evno, sysarg_t imethod)
  * @param evno    Event type to subscribe.
  * @param imethod Use this interface and method for notifying me.
  *
- * @return Value returned by the kernel.
+ * @return Error code returned by the kernel.
  *
  */
 int ipc_event_task_subscribe(event_task_type_t evno, sysarg_t imethod)
 {
-	return __SYSCALL2(SYS_IPC_EVENT_SUBSCRIBE, (sysarg_t) evno,
+	return (int) __SYSCALL2(SYS_IPC_EVENT_SUBSCRIBE, (sysarg_t) evno,
 	    (sysarg_t) imethod);
 }
 
@@ -71,48 +71,48 @@ int ipc_event_task_subscribe(event_task_type_t evno, sysarg_t imethod)
  *
  * @param evno    Event type to unsubscribe.
  *
- * @return Value returned by the kernel.
+ * @return Error code returned by the kernel.
  *
  */
 int ipc_event_unsubscribe(event_type_t evno)
 {
-	return __SYSCALL1(SYS_IPC_EVENT_UNSUBSCRIBE, (sysarg_t) evno);
+	return (int) __SYSCALL1(SYS_IPC_EVENT_UNSUBSCRIBE, (sysarg_t) evno);
 }
 
 /** Unsubscribe from task event notifications.
  *
  * @param evno    Event type to unsubscribe.
  *
- * @return Value returned by the kernel.
+ * @return Error code returned by the kernel.
  *
  */
 int ipc_event_task_unsubscribe(event_task_type_t evno)
 {
-	return __SYSCALL1(SYS_IPC_EVENT_UNSUBSCRIBE, (sysarg_t) evno);
+	return (int) __SYSCALL1(SYS_IPC_EVENT_UNSUBSCRIBE, (sysarg_t) evno);
 }
 
 /** Unmask event notifications.
  *
  * @param evno Event type to unmask.
  *
- * @return Value returned by the kernel.
+ * @return Error code returned by the kernel.
  *
  */
 int ipc_event_unmask(event_type_t evno)
 {
-	return __SYSCALL1(SYS_IPC_EVENT_UNMASK, (sysarg_t) evno);
+	return (int) __SYSCALL1(SYS_IPC_EVENT_UNMASK, (sysarg_t) evno);
 }
 
 /** Unmask task event notifications.
  *
  * @param evno Event type to unmask.
  *
- * @return Value returned by the kernel.
+ * @return Error code returned by the kernel.
  *
  */
 int ipc_event_task_unmask(event_task_type_t evno)
 {
-	return __SYSCALL1(SYS_IPC_EVENT_UNMASK, (sysarg_t) evno);
+	return (int) __SYSCALL1(SYS_IPC_EVENT_UNMASK, (sysarg_t) evno);
 }
 
 /** @}
